+++ foxy_farmer/cmds/join_pool.py
@@ -2,6 +2,7 @@ import asyncio
 import functools
 import time
 from asyncio import sleep, create_task
+from datetime import datetime
 from pathlib import Path
 from typing import Dict, Any, List, Optional
 
@@ -19,6 +20,7 @@ from chia.util.ints import uint64
 from chia.wallet.util.wallet_types import WalletType
 from chia.wallet.wallet_node import WalletNode
 from chia.wallet.wallet_node_api import WalletNodeAPI
+from humanize import naturaldelta
 from yaspin import yaspin
 
 from foxy_farmer.chia_launcher import ChiaLauncher
@@ -154,13 +156,28 @@ async def join_plot_nft_to_pool(wallet_client: WalletRpcClient, pool_info: Dict[
 
 
 async def wait_for_wallet_sync(wallet_client: WalletRpcClient):
-    with yaspin(text="Waiting for the wallet to sync ..."):
+    with yaspin(text="Waiting for the wallet to sync ...") as spinner:
+        async def update_spinner_text():
+            connected_full_nodes_count = len(await wallet_client.get_connections(node_type=NodeType.FULL_NODE))
+            wallet_height = await wallet_client.get_height_info()
+            relative_time = "N/A"
+            try:
+                wallet_timestamp = await wallet_client.get_timestamp_for_height(wallet_height)
+                relative_time = naturaldelta(datetime.now() - datetime.fromtimestamp(float(wallet_timestamp)))
+            except:
+                pass
+            spinner.text = f"Waiting for the wallet to sync (peers={connected_full_nodes_count}, height={wallet_height}, {relative_time} behind) ..."
+
         while len(await wallet_client.get_connections(node_type=NodeType.FULL_NODE)) < 2:
+            await update_spinner_text()
             await sleep(5)
+        await update_spinner_text()
         await sleep(10)
         while await wallet_client.get_sync_status():
+            await update_spinner_text()
             await sleep(5)
         while not (await wallet_client.get_synced()):
+            await update_spinner_text()
             await sleep(5)
     print("✅ Wallet synced")